Mengubah int menjadi long

+++ Praktikum 3/src/DiskonMiniMarket.java
@@ -5,13 +5,13 @@ public class DiskonMiniMarket {
         Scanner sc = new Scanner(System.in);
         System.out.println("Total pembelian: ");
         // tambahkan code untuk menyimpan nominal total pembelian
-        double total_pembelian = sc.nextInt();
+        long total_pembelian = sc.nextInt();
         // tambahkan code untuk mengimplementasikan ketentuan diskon yang ditetapkan
         if (total_pembelian >= 50000 && total_pembelian <= 75000)
-            total_pembelian = 0.95 * total_pembelian;
+            total_pembelian = (long) (0.95 * total_pembelian);
         else if (total_pembelian >= 75000 && total_pembelian <= 125000)
-            total_pembelian = 0.85 * total_pembelian;
-        else total_pembelian = 0.8 * total_pembelian;
+            total_pembelian = (long) (0.85 * total_pembelian);
+        else total_pembelian = (long) (0.8 * total_pembelian);
         // tambahkan code untuk menampilkan nominal yang harus dibayar sesuai dengan ketentuan diskon yang diberikan
         System.out.println("Total pembelian setelah diskon: " + total_pembelian);
     }
